requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18

faster-whisper==1.1.0
torch==2.5.1
torchaudio==2.5.1

pyannote.audio==3.3.2

python-dotenv==1.0.1
pydub==0.25.1
requests==2.31.0

#silero-vad==0.1.0        # VAD для chunking (когда дойдём)
#webvtt-py==1.0.0         # если потом захочешь VTT/SRT
